Refactoered into full stack component. Error handling.
